D10/Kafka_Flask/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Faker==18.7.0
Flask==2.3.2
Flask-HTTPAuth==4.8.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
python-dateutil==2.8.2
requests==2.30.0
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
